Update the ETL to use Spark SQL. Add more file formats.
